Fibonacci Sequence
Finding Fibonacci Sequence by first getting user to enter the first and second number, then asking user for the position in which they want the value of. The program will then add each number to each other consecutively until it reaches that position, then it outputs the value to the screen.